DiscordCompanion/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.1",
    "requests>=2.32.4",
    "flask-migrate>=4.1.0",
    "psycopg2-binary>=2.9.10",
    "asyncpg>=0.30.0",
    "sqlalchemy>=2.0.41",
]
